test: return-tag [rename from yield-tag]
