+++ packages/translator/src/__tests__/fixtures/return-tag/template.marko
@@ -1,5 +1,4 @@
-<attrs/{ show }/>
-<if=show>
+<if=input.show>
   <return=1/>
 </if>
 <else>
