+++ packages/translator/src/__tests__/fixtures/return-tag/test.ts
@@ -1,0 +1,1 @@
+export const steps = [{ show: true }];
