--- packages/translator/src/__tests__/fixtures/yield-tag/test.ts
@@ -1,2 +1,0 @@
-export const skip_csr = true;
-export const skip_ssr = true;
